variável valormensalidade do MODEL UNIPÊ alterada para DOUBLE

+++ src/main/java/dev/fujioka/java/avancado/web/model/Unipe.java
@@ -20,7 +20,7 @@ public class Unipe {
 
     private String curso;
 
-    private BigDecimal valorMensalidade;
+    private double valorMensalidade;
